test(rsc): share starter CSS HMR coverage
