+++ packages/plugin-rsc/e2e/css-link-precedence.test.ts
@@ -1,6 +1,5 @@
-import { expect, test } from '@playwright/test'
+import { test } from '@playwright/test'
 import { setupInlineFixture, useFixture } from './fixture'
-import { expectNoReload, waitForHydration } from './helper'
 import { defineStarterTest } from './starter'
 
 test.describe('cssLinkPrecedence-false', () => {
@@ -31,26 +30,6 @@ test.describe('cssLinkPrecedence-false', () => {
   test.describe('dev', () => {
     const f = useFixture({ root, mode: 'dev' })
     defineStarterTest(f)
-
-    // TODO: move css hmr test to `starter.ts`
-    test('css hmr', async ({ page }) => {
-      await page.goto(f.url())
-      await waitForHydration(page)
-      const card = page.locator('.card').nth(0)
-
-      await using _ = await expectNoReload(page)
-      const editor = f.createEditor('src/index.css')
-      editor.edit((s) =>
-        s.replace(
-          '.card {\n  padding: 1rem;',
-          `.card {\n  padding: 1rem; background-color: rgb(255, 0, 200);`,
-        ),
-      )
-      await expect(card).toHaveCSS('background-color', 'rgb(255, 0, 200)')
-
-      editor.reset()
-      await expect(card).not.toHaveCSS('background-color', 'rgb(255, 0, 200)')
-    })
   })
 
   test.describe('build', () => {
